Add config folder
